split tests for tokens
